backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymysql==1.1.1
pydantic==2.8.2
kafka-python==2.0.2
python-multipart==0.0.9
insightface==0.7.3
opencv-python==4.10.0.84
numpy==1.26.4